requirements.txt
gymnasium==0.29.1
pytorch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
pyyaml==6.0.1
pygame==2.2.0
python-flint==0.5.0
matplotlib==3.8.4
scipy==1.13.0